move codes to python
