+++ python/coral_usb/__init__.py
@@ -1,1 +1,4 @@
+from .detector_base import EdgeTPUDetectorBase  # NOQA
+from .face_detector import EdgeTPUFaceDetector  # NOQA
+from .object_detector import EdgeTPUObjectDetector  # NOQA
 from .posenet.pose_engine import PoseEngine  # NOQA
